Add basic execution tracing support to apf_run
This adds a --trace option to apf_run which dumps the PC and register
content during execution. This is a big aid in debugging misbehaving
APF programs.

It should be easy to disassemble the current instruction inline if we
refactor apf_disassembler to extract its body to a separate function.
Another easy feature would be dumping the memory locations that were
modified by the previous instruction.

Test: apf_run --trace --program <program> --packet <packet> --data <data>
Change-Id: I2a24bdfb84ce4e86409542332ee32082988a7698