Fix apf_run disassembler output.
The change to print disassembled code in apf_run (aosp/1129074)
broke the output of apf_run by printing PC, R0, PC instead of
R0, R1, PC.

This was noticed because compiling the code with -O0 (which
happened when I put apf_run.c into a test target) results in a
compiler error. Unfortunately the error goes away with the
default compiler settings and -Werror=format-extra-args does not
help.

Test: existing tests were incorrect, modified
Change-Id: Iad1713ce253b5df62ad5cd42b6ddff28559dda8c