Add test coverage for the --age parameter to apf_run.
Currently this is untested, and, as it happened, didn't work when
all the other parameters (program, packet, data, --trace) were
specified.

Currently each testcase is a set of 3 or 4 files with the same
basename and with extensions .program, .packet, .output, and
optionally .data. For simplicity, add another optional file with
extension .age.

Instead of recreating more testdata, re-use the existing testcase
one_ra_with_counters, adding three more variants of it, one with
age 30 (drop), one with age 600 (drop) and one with age 601
(pass). Avoid adding identical files by using symlinks for the
program, packet and data files, and only define new files for the
output and age files.

Test: atest apf_run_test
Change-Id: Idab8be57b1f98bd0246d9ec31695becce43c0d05