Add data_bins to apf_run_test so it can run remotely.
Currently it's possible to run apf_run_test locally, but not
remotely, because the required binary (apf_run) is not packaged
when the test is built.

Fix this using the new data_bins property, and remove the useless
requires: which was an attempt to fix this problem. Make the
test shell script invoke apf_run with "./" because the binary is
in the same directory as the test (it used to work when running
locally because it was in a directory that was in $PATH).

Also annotate the test with a null-device property, since it does
not require a device. This does not seem to work with atest but
hopefully it does work when running remotely.

Bug: 143248826
Test: atest apf_run_test
Test: forrest run L17500000674511176
Change-Id: Ifec451c5ab4f4c96ad91a9a64157d88800e443e2