v5: replace PASS_PACKET with EXCEPTION
PASS_PACKET should be replaced with PASS as it is needlessly
verbose - what else would we be passing?

However, the majority of the cases where we return PASS_PACKET
are actually EXCEPTIONs.  As such we s/PASS_PACKET/EXCEPTION/,
and we'll follow up with s/EXCEPTION/PASS/ on only the *normal*
cases in the next commit.

Generated via:
  sed -r -i 's@PASS_PACKET@EXCEPTION@g' apf_interpreter_source.c

Test: TreeHugger
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: If6496deb62c6070862c8cfc93f594543604b22ba